Rename uploadRouter to moodRouter
